refactor: unify portal promo codes service
