+++ API_DOCUMENTATION.md
@@ -471,6 +471,9 @@ Response 200:
 - `POST /portal/promocodes/issue` — создание промокода (см. `PortalPromoCodePayload`).
 - `PUT /portal/promocodes/:promoCodeId` — обновление.
 - `POST /portal/promocodes/deactivate` / `POST /portal/promocodes/activate` — смена статуса.
+- `GET /portal/loyalty/promocodes?status=ACTIVE|ARCHIVE|ALL` — доступ к «сырым» полям `PromoCode` (включая метрики) для страницы лояльности.
+- `POST /portal/loyalty/promocodes` — создание через `LoyaltyPromoCodePayload` (код, сегмент, уровни, лимиты, autoArchive).
+- `PUT /portal/loyalty/promocodes/:id` / `POST /portal/loyalty/promocodes/:id/status` / `POST /portal/loyalty/promocodes/bulk/status` — управление состояниями и массовое архивирование.
 
 В API лояльности промокод передаётся полем `promoCode` в `POST /loyalty/quote` и `POST /loyalty/commit`. При применении начисляются дополнительные баллы, TTL и уровень согласно настройкам.
