+++ api/src/loyalty-program/loyalty-program.service.ts
@@ -1,13 +1,5 @@
 import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
-import {
-  LoyaltyMechanicType,
-  MechanicStatus,
-  Prisma,
-  PromotionRewardType,
-  PromotionStatus,
-  PromoCodeStatus,
-  PromoCodeUsageLimitType,
-} from '@prisma/client';
+import { LoyaltyMechanicType, MechanicStatus, Prisma, PromotionRewardType, PromotionStatus } from '@prisma/client';
 import { PrismaService } from '../prisma.service';
 import { MetricsService } from '../metrics.service';
 
@@ -74,30 +66,6 @@ export interface PromotionPayload {
   actorId?: string;
 }
 
-export interface PromoCodePayload {
-  code: string;
-  name?: string | null;
-  description?: string | null;
-  status?: PromoCodeStatus;
-  segmentId?: string | null;
-  usageLimitType?: PromoCodeUsageLimitType;
-  usageLimitValue?: number | null;
-  cooldownDays?: number | null;
-  perCustomerLimit?: number | null;
-  requireVisit?: boolean;
-  visitLookbackHours?: number | null;
-  grantPoints?: boolean;
-  pointsAmount?: number | null;
-  pointsExpireInDays?: number | null;
-  assignTierId?: string | null;
-  upgradeTierId?: string | null;
-  activeFrom?: Date | string | null;
-  activeUntil?: Date | string | null;
-  autoArchiveAt?: Date | string | null;
-  isHighlighted?: boolean;
-  metadata?: any;
-  actorId?: string;
-}
 
 export interface OperationsLogFilters {
   type?: 'MECHANIC' | 'PROMO_CODE' | 'PROMOTION';
@@ -625,181 +593,6 @@ export class LoyaltyProgramService {
     return { updated };
   }
 
-  async listPromoCodes(merchantId: string, status?: PromoCodeStatus | 'ALL') {
-    const where: Prisma.PromoCodeWhereInput = { merchantId };
-    if (status && status !== 'ALL') where.status = status;
-
-    const promoCodes = await this.prisma.promoCode.findMany({
-      where,
-      orderBy: { createdAt: 'desc' },
-      include: { metrics: true },
-    });
-
-    try {
-      this.logger.log(
-        JSON.stringify({
-          event: 'portal.loyalty.promocodes.list',
-          merchantId,
-          status: status ?? 'ALL',
-          total: promoCodes.length,
-        }),
-      );
-      this.metrics.inc('portal_loyalty_promocodes_list_total');
-    } catch {}
-    return promoCodes;
-  }
-
-  async createPromoCode(merchantId: string, payload: PromoCodePayload) {
-    if (!payload.code?.trim()) throw new BadRequestException('Код обязателен');
-    const exists = await this.prisma.promoCode.findFirst({ where: { merchantId, code: payload.code.trim() } });
-    if (exists) throw new BadRequestException('Промокод уже существует');
-
-    const promoCode = await this.prisma.promoCode.create({
-      data: {
-        merchantId,
-        code: payload.code.trim(),
-        name: payload.name ?? null,
-        description: payload.description ?? null,
-        status: payload.status ?? PromoCodeStatus.DRAFT,
-        segmentId: payload.segmentId ?? null,
-        usageLimitType: payload.usageLimitType ?? PromoCodeUsageLimitType.UNLIMITED,
-        usageLimitValue: payload.usageLimitValue ?? null,
-        cooldownDays: payload.cooldownDays ?? null,
-        perCustomerLimit: payload.perCustomerLimit ?? null,
-        requireVisit: payload.requireVisit ?? false,
-        visitLookbackHours: payload.visitLookbackHours ?? null,
-        grantPoints: payload.grantPoints ?? false,
-        pointsAmount: payload.pointsAmount ?? null,
-        pointsExpireInDays: payload.pointsExpireInDays ?? null,
-        assignTierId: payload.assignTierId ?? null,
-        upgradeTierId: payload.upgradeTierId ?? null,
-        activeFrom: payload.activeFrom ? new Date(payload.activeFrom) : null,
-        activeUntil: payload.activeUntil ? new Date(payload.activeUntil) : null,
-        autoArchiveAt: payload.autoArchiveAt ? new Date(payload.autoArchiveAt) : null,
-        isHighlighted: payload.isHighlighted ?? false,
-        metadata: payload.metadata ?? null,
-        createdById: payload.actorId ?? null,
-        updatedById: payload.actorId ?? null,
-      },
-    });
-
-    try {
-      this.logger.log(
-        JSON.stringify({
-          event: 'portal.loyalty.promocodes.create',
-          merchantId,
-          promoCodeId: promoCode.id,
-          status: promoCode.status,
-        }),
-      );
-      this.metrics.inc('portal_loyalty_promocodes_changed_total', { action: 'create' });
-    } catch {}
-    return promoCode;
-  }
-
-  async updatePromoCode(merchantId: string, promoCodeId: string, payload: PromoCodePayload) {
-    const promoCode = await this.prisma.promoCode.findFirst({ where: { merchantId, id: promoCodeId } });
-    if (!promoCode) throw new NotFoundException('Промокод не найден');
-
-    const updated = await this.prisma.promoCode.update({
-      where: { id: promoCodeId },
-      data: {
-        code: payload.code?.trim() ?? promoCode.code,
-        name: payload.name ?? promoCode.name,
-        description: payload.description ?? promoCode.description,
-        status: payload.status ?? promoCode.status,
-        segmentId: payload.segmentId ?? promoCode.segmentId,
-        usageLimitType: payload.usageLimitType ?? promoCode.usageLimitType,
-        usageLimitValue: payload.usageLimitValue ?? promoCode.usageLimitValue,
-        cooldownDays: payload.cooldownDays ?? promoCode.cooldownDays,
-        perCustomerLimit: payload.perCustomerLimit ?? promoCode.perCustomerLimit,
-        requireVisit: payload.requireVisit ?? promoCode.requireVisit,
-        visitLookbackHours: payload.visitLookbackHours ?? promoCode.visitLookbackHours,
-        grantPoints: payload.grantPoints ?? promoCode.grantPoints,
-        pointsAmount: payload.pointsAmount ?? promoCode.pointsAmount,
-        pointsExpireInDays: payload.pointsExpireInDays ?? promoCode.pointsExpireInDays,
-        assignTierId: payload.assignTierId ?? promoCode.assignTierId,
-        upgradeTierId: payload.upgradeTierId ?? promoCode.upgradeTierId,
-        activeFrom: payload.activeFrom ? new Date(payload.activeFrom) : promoCode.activeFrom,
-        activeUntil: payload.activeUntil ? new Date(payload.activeUntil) : promoCode.activeUntil,
-        autoArchiveAt: payload.autoArchiveAt ? new Date(payload.autoArchiveAt) : promoCode.autoArchiveAt,
-        isHighlighted: payload.isHighlighted ?? promoCode.isHighlighted,
-        metadata: payload.metadata ?? promoCode.metadata,
-        updatedById: payload.actorId ?? promoCode.updatedById,
-      },
-    });
-
-    try {
-      this.logger.log(
-        JSON.stringify({
-          event: 'portal.loyalty.promocodes.update',
-          merchantId,
-          promoCodeId,
-          status: updated.status,
-        }),
-      );
-      this.metrics.inc('portal_loyalty_promocodes_changed_total', { action: 'update' });
-    } catch {}
-    return updated;
-  }
-
-  async changePromoCodeStatus(merchantId: string, promoCodeId: string, status: PromoCodeStatus, actorId?: string) {
-    const promoCode = await this.prisma.promoCode.findFirst({ where: { merchantId, id: promoCodeId } });
-    if (!promoCode) throw new NotFoundException('Промокод не найден');
-
-    const updated = await this.prisma.promoCode.update({
-      where: { id: promoCodeId },
-      data: {
-        status,
-        updatedById: actorId ?? promoCode.updatedById,
-        archivedAt: status === PromoCodeStatus.ARCHIVED ? new Date() : promoCode.archivedAt,
-      },
-    });
-
-    try {
-      this.logger.log(
-        JSON.stringify({
-          event: 'portal.loyalty.promocodes.status',
-          merchantId,
-          promoCodeId,
-          status,
-        }),
-      );
-      this.metrics.inc('portal_loyalty_promocodes_changed_total', { action: 'status' });
-    } catch {}
-    return updated;
-  }
-
-  async bulkArchivePromoCodes(merchantId: string, promoCodeIds: string[], status: PromoCodeStatus, actorId?: string) {
-    if (!promoCodeIds.length) return { updated: 0 };
-    const results = await this.prisma.$transaction(
-      promoCodeIds.map((id) =>
-        this.prisma.promoCode.updateMany({
-          where: { id, merchantId },
-          data: {
-            status,
-            updatedById: actorId ?? undefined,
-            archivedAt: status === PromoCodeStatus.ARCHIVED ? new Date() : undefined,
-          },
-        }),
-      ),
-    );
-
-    const updated = results.reduce((acc, res) => acc + res.count, 0);
-    try {
-      this.logger.log(
-        JSON.stringify({
-          event: 'portal.loyalty.promocodes.bulkStatus',
-          merchantId,
-          status,
-          ids: promoCodeIds.length,
-          updated,
-        }),
-      );
-      this.metrics.inc('portal_loyalty_promocodes_changed_total', { action: 'bulk-status' }, updated || 1);
-    } catch {}
-    return { updated };
-  }
 
   async operationsLog(merchantId: string, filters: OperationsLogFilters = {}) {
     const from = filters.from ? new Date(filters.from) : undefined;
