+++ plan.md
@@ -283,6 +283,7 @@
 — Следующий шаг (Wave 2, текущая сессия):
   - Cashier UI: реализовать экран кассира (login 9‑значный пароль мерчанта → ввод PIN сотрудника по точке) и привязку к `X-Staff-Key`.
   - Promocodes (points): расширить обработку в `LoyaltyService.commit()` для начислений по промокоду (POINTS) — idempotent по `orderId`, структурные логи/метрики; e2e.
+    - [x] Централизовать бизнес-логику промокодов в `PromoCodesService`, перевести портал/loyalty-контроллеры и метрики на единый слой.
   - E2E: добавить кейс earn по промокоду POINTS (issue → quote/commit с promoCode → проверка баланса/транзакций).
 
 ## Волна 3 — Завершена (2025-09-15)
